better documentation and tests
