+++ tests/web/rpc/v1/test_endpoint_trace_item_table/test_endpoint_trace_item_table.py
@@ -53,7 +53,6 @@ from snuba.web.rpc.common.exceptions import BadSnubaRPCRequestException
 from snuba.web.rpc.v1.endpoint_trace_item_table import (
     EndpointTraceItemTable,
     _apply_labels_to_columns,
-    convert_to_conditional_aggregation,
 )
 from tests.base import BaseApiTest
 from tests.helpers import write_raw_unprocessed_events
@@ -2785,67 +2784,6 @@ class TestTraceItemTable(BaseApiTest):
         ]
 
 
-def _build_sum_attribute_aggregation_column_with_name(name: str) -> Column:
-    return Column(
-        aggregation=AttributeAggregation(
-            aggregate=Function.FUNCTION_SUM,
-            key=AttributeKey(type=AttributeKey.TYPE_DOUBLE, name=name),
-            label="sum(" + name + ")",
-            extrapolation_mode=ExtrapolationMode.EXTRAPOLATION_MODE_NONE,
-        )
-    )
-
-
-def _build_sum_attribute_conditional_aggregation_column_with_name(name: str) -> Column:
-    return Column(
-        conditional_aggregation=AttributeConditionalAggregation(
-            aggregate=Function.FUNCTION_SUM,
-            key=AttributeKey(type=AttributeKey.TYPE_DOUBLE, name=name),
-            label="sum(" + name + ")",
-            extrapolation_mode=ExtrapolationMode.EXTRAPOLATION_MODE_NONE,
-        ),
-    )
-
-
-def _build_avg_aggregation_comparison_filter_with_name(
-    name: str,
-) -> AggregationComparisonFilter:
-    return AggregationComparisonFilter(
-        aggregation=AttributeAggregation(
-            aggregate=Function.FUNCTION_AVG,
-            key=AttributeKey(
-                type=AttributeKey.TYPE_DOUBLE,
-                name=name,
-            ),
-        ),
-        op=AggregationComparisonFilter.OP_GREATER_THAN,
-        val=350,
-    )
-
-
-def _build_avg_conditional_aggregation_comparison_filter_with_name(
-    name: str,
-) -> AggregationComparisonFilter:
-    return AggregationComparisonFilter(
-        aggregation=AttributeAggregation(
-            aggregate=Function.FUNCTION_AVG,
-            key=AttributeKey(
-                type=AttributeKey.TYPE_DOUBLE,
-                name=name,
-            ),
-        ),
-        conditional_aggregation=AttributeConditionalAggregation(
-            aggregate=Function.FUNCTION_AVG,
-            key=AttributeKey(
-                type=AttributeKey.TYPE_DOUBLE,
-                name=name,
-            ),
-        ),
-        op=AggregationComparisonFilter.OP_GREATER_THAN,
-        val=350,
-    )
-
-
 class TestUtils:
     def test_apply_labels_to_columns_backward_compat(self) -> None:
         message = TraceItemTableRequest(
@@ -2910,211 +2848,3 @@ class TestUtils:
         _apply_labels_to_columns(message)
         assert message.columns[0].label == "avg(custom_measurement)"
         assert message.columns[1].label == "avg(custom_measurement_2)"
-
-    def test_convert_aggregation_to_conditional_aggregation(self) -> None:
-        message = TraceItemTableRequest(
-            meta=RequestMeta(
-                project_ids=[1, 2, 3],
-                organization_id=1,
-                cogs_category="something",
-                referrer="something",
-                start_timestamp=Timestamp(seconds=1),
-                end_timestamp=Timestamp(seconds=2),
-                trace_item_type=TraceItemType.TRACE_ITEM_TYPE_SPAN,
-            ),
-            columns=[
-                Column(
-                    key=AttributeKey(type=AttributeKey.TYPE_STRING, name="doesntmatter")
-                ),
-                _build_sum_attribute_aggregation_column_with_name("column_1"),
-                _build_sum_attribute_aggregation_column_with_name("column_2"),
-                Column(
-                    formula=Column.BinaryFormula(
-                        op=Column.BinaryFormula.OP_DIVIDE,
-                        left=Column(
-                            formula=Column.BinaryFormula(
-                                op=Column.BinaryFormula.OP_MULTIPLY,
-                                left=Column(
-                                    formula=Column.BinaryFormula(
-                                        op=Column.BinaryFormula.OP_ADD,
-                                        left=_build_sum_attribute_aggregation_column_with_name(
-                                            "add_1"
-                                        ),
-                                        right=_build_sum_attribute_aggregation_column_with_name(
-                                            "add_2"
-                                        ),
-                                    ),
-                                    label="mult_1",
-                                ),
-                                right=_build_sum_attribute_aggregation_column_with_name(
-                                    "mult_2"
-                                ),
-                            ),
-                            label="divide_1",
-                        ),
-                        right=_build_sum_attribute_aggregation_column_with_name(
-                            "divide_2"
-                        ),
-                    ),
-                    label="((sum(add_1) + sum(add_2)) * sum(mult_2)) / sum(divide_2)",
-                ),
-            ],
-            order_by=[
-                TraceItemTableRequest.OrderBy(
-                    column=_build_sum_attribute_aggregation_column_with_name("column_1")
-                ),
-                TraceItemTableRequest.OrderBy(
-                    column=Column(
-                        formula=Column.BinaryFormula(
-                            op=Column.BinaryFormula.OP_DIVIDE,
-                            left=_build_sum_attribute_aggregation_column_with_name(
-                                "column_1"
-                            ),
-                            right=_build_sum_attribute_aggregation_column_with_name(
-                                "column_2"
-                            ),
-                        ),
-                        label="sum(column_1) / sum(column_2)",
-                    )
-                ),
-            ],
-            aggregation_filter=AggregationFilter(  # same filter on both sides of the and
-                and_filter=AggregationAndFilter(
-                    filters=[
-                        AggregationFilter(
-                            comparison_filter=_build_avg_aggregation_comparison_filter_with_name(
-                                "column_3"
-                            )
-                        ),
-                        AggregationFilter(
-                            and_filter=AggregationAndFilter(
-                                filters=[
-                                    AggregationFilter(
-                                        or_filter=AggregationOrFilter(
-                                            filters=[
-                                                AggregationFilter(
-                                                    comparison_filter=_build_avg_aggregation_comparison_filter_with_name(
-                                                        "column_4"
-                                                    )
-                                                ),
-                                                AggregationFilter(
-                                                    comparison_filter=_build_avg_aggregation_comparison_filter_with_name(
-                                                        "column_5"
-                                                    )
-                                                ),
-                                            ]
-                                        )
-                                    ),
-                                    AggregationFilter(
-                                        comparison_filter=_build_avg_aggregation_comparison_filter_with_name(
-                                            "column_6"
-                                        )
-                                    ),
-                                ]
-                            )
-                        ),
-                    ]
-                )
-            ),
-        )
-
-        convert_to_conditional_aggregation(message)
-        assert message.columns == [
-            Column(
-                key=AttributeKey(type=AttributeKey.TYPE_STRING, name="doesntmatter")
-            ),
-            _build_sum_attribute_conditional_aggregation_column_with_name("column_1"),
-            _build_sum_attribute_conditional_aggregation_column_with_name("column_2"),
-            Column(
-                formula=Column.BinaryFormula(
-                    op=Column.BinaryFormula.OP_DIVIDE,
-                    left=Column(
-                        formula=Column.BinaryFormula(
-                            op=Column.BinaryFormula.OP_MULTIPLY,
-                            left=Column(
-                                formula=Column.BinaryFormula(
-                                    op=Column.BinaryFormula.OP_ADD,
-                                    left=_build_sum_attribute_conditional_aggregation_column_with_name(
-                                        "add_1"
-                                    ),
-                                    right=_build_sum_attribute_conditional_aggregation_column_with_name(
-                                        "add_2"
-                                    ),
-                                ),
-                                label="mult_1",
-                            ),
-                            right=_build_sum_attribute_conditional_aggregation_column_with_name(
-                                "mult_2"
-                            ),
-                        ),
-                        label="divide_1",
-                    ),
-                    right=_build_sum_attribute_conditional_aggregation_column_with_name(
-                        "divide_2"
-                    ),
-                ),
-                label="((sum(add_1) + sum(add_2)) * sum(mult_2)) / sum(divide_2)",
-            ),
-        ]
-        assert message.order_by == [
-            TraceItemTableRequest.OrderBy(
-                column=_build_sum_attribute_conditional_aggregation_column_with_name(
-                    "column_1"
-                )
-            ),
-            TraceItemTableRequest.OrderBy(
-                column=Column(
-                    formula=Column.BinaryFormula(
-                        op=Column.BinaryFormula.OP_DIVIDE,
-                        left=_build_sum_attribute_conditional_aggregation_column_with_name(
-                            "column_1"
-                        ),
-                        right=_build_sum_attribute_conditional_aggregation_column_with_name(
-                            "column_2"
-                        ),
-                    ),
-                    label="sum(column_1) / sum(column_2)",
-                )
-            ),
-        ]
-        assert (
-            message.aggregation_filter
-            == AggregationFilter(  # same filter on both sides of the and
-                and_filter=AggregationAndFilter(
-                    filters=[
-                        AggregationFilter(
-                            comparison_filter=_build_avg_conditional_aggregation_comparison_filter_with_name(
-                                "column_3"
-                            )
-                        ),
-                        AggregationFilter(
-                            and_filter=AggregationAndFilter(
-                                filters=[
-                                    AggregationFilter(
-                                        or_filter=AggregationOrFilter(
-                                            filters=[
-                                                AggregationFilter(
-                                                    comparison_filter=_build_avg_conditional_aggregation_comparison_filter_with_name(
-                                                        "column_4"
-                                                    )
-                                                ),
-                                                AggregationFilter(
-                                                    comparison_filter=_build_avg_conditional_aggregation_comparison_filter_with_name(
-                                                        "column_5"
-                                                    )
-                                                ),
-                                            ]
-                                        )
-                                    ),
-                                    AggregationFilter(
-                                        comparison_filter=_build_avg_conditional_aggregation_comparison_filter_with_name(
-                                            "column_6"
-                                        )
-                                    ),
-                                ]
-                            )
-                        ),
-                    ]
-                )
-            )
-        )
